+++ src/DownKyi/ViewModels/ViewMyBangumiFollowViewModel.cs
@@ -152,8 +152,8 @@ namespace DownKyi.ViewModels
 
             TabHeaders = new ObservableCollection<TabHeader>
             {
-                new TabHeader { Id = (int)Core.BiliApi.Users.Models.BangumiType.ANIME, Title = DictionaryResource.GetString("FollowAnime") },
-                new TabHeader { Id = (int)Core.BiliApi.Users.Models.BangumiType.EPISODE, Title = DictionaryResource.GetString("FollowMovie") }
+                new TabHeader { Id = (long)Core.BiliApi.Users.Models.BangumiType.ANIME, Title = DictionaryResource.GetString("FollowAnime") },
+                new TabHeader { Id = (long)Core.BiliApi.Users.Models.BangumiType.EPISODE, Title = DictionaryResource.GetString("FollowMovie") }
             };
 
             Medias = new ObservableCollection<BangumiFollowMedia>();
